Fix test suite
